minnet re-execute hack with checkpoints

+++ cmd/hack/hack.go
@@ -26,6 +26,7 @@ import (
 	"github.com/ledgerwatch/bolt"
 	"github.com/ledgerwatch/turbo-geth/common"
 	"github.com/ledgerwatch/turbo-geth/common/dbutils"
+	"github.com/ledgerwatch/turbo-geth/common/debug"
 	"github.com/ledgerwatch/turbo-geth/consensus/ethash"
 	"github.com/ledgerwatch/turbo-geth/core"
 	"github.com/ledgerwatch/turbo-geth/core/rawdb"
@@ -613,37 +614,52 @@ func trieChart() {
 	check(err)
 }
 
-func execToBlock(block int) {
-	blockDb, err := ethdb.NewBoltDatabase(node.DefaultDataDir() + "/geth-remove-me/geth/chaindata")
+func execToBlock(block uint64, fromScratch bool) {
+	blockDb, err := ethdb.NewBoltDatabase(node.DefaultDataDir() + "/geth/chaindata")
 	//ethDb, err := ethdb.NewBoltDatabase("/home/akhounov/.ethereum/geth/chaindata")
 	check(err)
-	bcb, err := core.NewBlockChain(blockDb, nil, params.TestnetChainConfig, ethash.NewFaker(), vm.Config{}, nil)
+	bcb, err := core.NewBlockChain(blockDb, nil, params.MainnetChainConfig, ethash.NewFaker(), vm.Config{}, nil)
 	check(err)
 	defer blockDb.Close()
-	os.Remove("statedb")
+	if fromScratch {
+		os.Remove("statedb")
+	}
 	stateDb, err := ethdb.NewBoltDatabase("statedb")
 	check(err)
 	defer stateDb.Close()
-	_, _, _, err = core.SetupGenesisBlock(stateDb, core.DefaultTestnetGenesisBlock())
+
+	_, _, _, err = core.SetupGenesisBlock(stateDb, core.DefaultGenesisBlock())
+	//_, _, _, err = core.SetupGenesisBlockWithOverride(stateDb, nil, nil, nil)
 	check(err)
-	bc, err := core.NewBlockChain(stateDb, nil, params.TestnetChainConfig, ethash.NewFaker(), vm.Config{}, nil)
+	bc, err := core.NewBlockChain(stateDb, nil, params.MainnetChainConfig, ethash.NewFaker(), vm.Config{}, nil)
 	check(err)
+	tds, err := bc.GetTrieDbState()
+	check(err)
+	if debug.IsIntermediateTrieHash() && fromScratch {
+		_ = tds.Rebuild()
+	}
+
+	importedBn := tds.GetBlockNr()
+
 	//bc.SetNoHistory(true)
 	blocks := types.Blocks{}
 	var lastBlock *types.Block
-	for i := 1; i <= block; i++ {
-		lastBlock = bcb.GetBlockByNumber(uint64(i))
+
+	now := time.Now()
+
+	for i := importedBn; i <= block; i++ {
+		lastBlock = bcb.GetBlockByNumber(i)
 		blocks = append(blocks, lastBlock)
 		if len(blocks) >= 100 || i == block {
 			_, err = bc.InsertChain(context.Background(), blocks)
-			check(err)
-			fmt.Printf("Inserted %d blocks\n", i)
+			if err != nil {
+				panic(err)
+			}
 			blocks = types.Blocks{}
 		}
-	}
-	tds, err := bc.GetTrieDbState()
-	if err != nil {
-		panic(err)
+		if i%1000 == 0 {
+			fmt.Printf("Inserted %dK, %s \n", i/1000, time.Since(now))
+		}
 	}
 	root := tds.LastRoot()
 	fmt.Printf("Root hash: %x\n", root)
@@ -773,6 +789,9 @@ func testStartup() {
 }
 
 func testResolveCached() {
+	execToBlock(3989604, true)
+	return
+
 	//startTime := time.Now()
 	ethDb, err := ethdb.NewBoltDatabase(node.DefaultDataDir() + "/geth-remove-me/geth/chaindata")
 	check(err)
